common constructor for image and audio data, remove unnecessary audio attribute
